% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digitraffic-package.R
\docType{package}
\name{digitraffic-package}
\alias{digitraffic}
\alias{digitraffic-package}
\title{digitraffic: R Client for the Finntraffic Digitraffic TMS (LAM) API}
\description{
Provides functions to access traffic measurement data from the Finnish Transport Infrastructure Agency's Digitraffic open API. Covers TMS (Traffic Measurement System, known as LAM in Finnish) stations including real-time sensor data, station metadata, sensor calibration constants, and historical raw vehicle-level observations. All data is returned as tidy tibbles ready for analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/theo-maon/digitraffic}
  \item Report bugs at \url{https://github.com/theo-maon/digitraffic/issues}
}

}
\author{
\strong{Maintainer}: Theo Maon \email{ma.oon.theo@gmail.com}

Authors:
\itemize{
  \item Theo Maon \email{ma.oon.theo@gmail.com}
}

}
\keyword{internal}
